Replaced parseInt with parseFloat in CreateCampaign.tsx so fractional budget/reward values (e.g., 0.01) are no longer truncated and compute correctly.

X-Lovable-Edit-ID: edt-29e13426-aa73-4070-b605-5220de4bde8f

+++ src/pages/CreateCampaign.tsx
@@ -88,8 +88,8 @@ const CreateCampaign = () => {
   const [reward, setReward] = useState("");
   const [done, setDone] = useState(false);
 
-  const budgetNum = parseInt(budget) || 0;
-  const rewardNum = parseInt(reward) || 1;
+  const budgetNum = parseFloat(budget) || 0;
+  const rewardNum = parseFloat(reward) || 1;
   const estimatedReach = rewardNum > 0 ? Math.floor(budgetNum / rewardNum) : 0;
   const canAfford = budgetNum <= credits && budgetNum > 0 && rewardNum > 0;
